% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cellAnno}
\alias{cellAnno}
\title{Annotation for single cell data at cluster level}
\usage{
cellAnno(query.srt, ref.tissue = "all", resolution = 0.5)
}
\arguments{
\item{query.srt}{query single cell data in Seurat object format}

\item{ref.tissue}{reference tissue type, default is "all". Other options include "immune" and specific tissue(s). To see list of available tissues, use list.tissues(). Multiple tissues separated by "|"}

\item{resolution}{resolution used to identify cell clusters, default is 0.5}
}
\value{
Seurat object with predicted.celltype added to metadata
}
\description{
Annotate single cell at cluster level using reference built from Tabula sapiens, human protein atlas and public datasets. The annotation algorithm used is SingleR.
}
