% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cooPredict}
\alias{cooPredict}
\title{Cell-of-Origin prediction for cancer bulk RNA-seq data}
\usage{
cooPredict(logtpm.matrix, ref = "core", exclude.tissues = NULL)
}
\arguments{
\item{logtpm.matrix}{query data matrix in log2-TPM}

\item{ref}{reference to use, default is "core".To see list of available tissues, use list.tissues(). Multiple tissues separated by "|". Set to "core" if not sure (for samples from metastatic sites, it is possible top hit is cell from the biopsy site).}
}
\value{
output of SingleR. "labels" is the predicted cell type.
}
\description{
Predict COO using reference built from Tabula sapiens, human protein atlas and public datasets. Right now it only support prediction for solid tumor type
}
