% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cellAnno}
\alias{cellAnno}
\title{Annotation for single cell data at cluster level}
\usage{
cellAnno(
  query.srt,
  ref.tissue = "all",
  cluster = NULL,
  resolution = 0.5,
  level2 = FALSE
)
}
\arguments{
\item{query.srt}{query single cell data in Seurat object format}

\item{ref.tissue}{reference tissue type, default is "all". Other options include "immune" and specific tissue(s). To see list of available tissues, use list.tissues(). Multiple tissues separated by "|"}

\item{cluster}{user provided cluster assignment to use. it should be a column name in meta.data. if NULL, will use Seurat::FindCluster to find clusters}

\item{resolution}{resolution used to identify cell clusters, Default is 0.5 if cluster is not set.}

\item{level2}{if to run an additional annotation step to get detailed cell subtypes. It only works when ref.tissue = "immune".}
}
\value{
Seurat object with "labels","pruned.labels" and "score" added to metadata
}
\description{
Annotate single cell at cluster level using reference built from Tabula sapiens, human protein atlas and public datasets. The annotation algorithm used is SingleR.
}
